refactor(core): move db properties to external folder
