+++ src/main/kotlin/org/careerseekers/csbackupmanager/config/properties/EventsServiceDatabaseConfigurationProperties.kt
@@ -1,11 +1,11 @@
-package org.careerseekers.csbackupmanager.config
+package org.careerseekers.csbackupmanager.config.properties
 
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.context.annotation.Configuration
 
 @Configuration
-@ConfigurationProperties(prefix = "database.fs-pg")
-class FilesServiceDatabaseConfigurationProperties {
+@ConfigurationProperties(prefix = "database.es-pg")
+class EventsServiceDatabaseConfigurationProperties {
     lateinit var url: String
     lateinit var username: String
     lateinit var password: String
