+++ src/main/kotlin/org/careerseekers/csbackupmanager/config/properties/FilesServiceDatabaseConfigurationProperties.kt
@@ -1,11 +1,11 @@
-package org.careerseekers.csbackupmanager.config
+package org.careerseekers.csbackupmanager.config.properties
 
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.context.annotation.Configuration
 
 @Configuration
-@ConfigurationProperties(prefix = "database.es-pg")
-class EventsServiceDatabaseConfigurationProperties {
+@ConfigurationProperties(prefix = "database.fs-pg")
+class FilesServiceDatabaseConfigurationProperties {
     lateinit var url: String
     lateinit var username: String
     lateinit var password: String
