+++ src/main/kotlin/org/careerseekers/csbackupmanager/config/properties/UserServiceDatabaseConfigurationProperties.kt
@@ -1,4 +1,4 @@
-package org.careerseekers.csbackupmanager.config
+package org.careerseekers.csbackupmanager.config.properties
 
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.context.annotation.Configuration
